+++ include/hicma/classes/initialization_helpers/matrix_initializer_block.h
@@ -30,7 +30,7 @@ class MatrixInitializerBlock : public MatrixInitializer {
   MatrixInitializerBlock& operator=(MatrixInitializerBlock&& A) = default;
 
   // Additional constructors
-  MatrixInitializerBlock(Dense&& A, int64_t admis, int64_t rank);
+  MatrixInitializerBlock(Dense&& A, double admis, int64_t rank);
 
   // Utility methods
   void fill_dense_representation(
