+++ include/hicma/classes/initialization_helpers/matrix_initializer_file.h
@@ -33,7 +33,7 @@ class MatrixInitializerFile : public MatrixInitializer {
 
   // Additional constructors
   MatrixInitializerFile(
-    std::string filename, int ordering, int64_t admis, int64_t rank, int basis_type
+    std::string filename, int ordering, double admis, int64_t rank, int basis_type
   );
 
   // Utility methods
