+++ include/hicma/classes/initialization_helpers/matrix_initializer_kernel.h
@@ -46,7 +46,7 @@ class MatrixInitializerKernel : public MatrixInitializer {
       const std::vector<std::vector<double>>& x,
       int64_t row_start, int64_t col_start
     ),
-    const std::vector<std::vector<double>>& x, int64_t admis, int64_t rank
+    const std::vector<std::vector<double>>& x, double admis, int64_t rank
   );
 
   // Utility methods
