+++ src/classes/initialization_helpers/matrix_initializer_file.cpp
@@ -15,7 +15,7 @@ namespace hicma
 
 // Additional constructors
 MatrixInitializerFile::MatrixInitializerFile(
-  std::string filename, int ordering, int64_t admis, int64_t rank, int basis_type
+  std::string filename, int ordering, double admis, int64_t rank, int basis_type
 ) : MatrixInitializer(admis, rank, basis_type), filename(filename), ordering(ordering) {}
 
 // Utility methods
